fix(review): 8 即修 + 10 issue (cron-cr v9 第六轮) (#2304)
* fix(cli): read shim error frame Msg field so error text is not lost [R202606f-ARCH-1]

* fix(node): recognize full 127.0.0.0/8 loopback range in isLoopbackHost [R202606f-SEC-002]

* perf(shim): skip session-id scan once known + hoist search patterns [R202606f-PERF-003]

* perf(node): use ReadMessage+Unmarshal in reverseconn readLoop to avoid per-frame decoder alloc [R202606f-PERF-001]

* fix(cli): discard pending slots in readLoop panic recover [R202606f-GO-008]

* fix(cli): deny --model in extra args to prevent model override [R202606f-SEC-001]

* fix(cli): store shim exit code as int64 to avoid 32-bit truncation [R202606f-GO-001]

* docs(cosmetic): cron-cr-20260622-200155 追加 9 行

---------

+++ internal/cli/process_readloop_codefield_test.go
@@ -20,7 +20,7 @@ func TestShimMsgCode_AbsentVsPresent(t *testing.T) {
 		name        string
 		raw         string
 		wantPresent bool
-		wantValue   int
+		wantValue   int64
 	}{
 		{name: "absent", raw: `{"type":"cli_exited"}`, wantPresent: false, wantValue: 0},
 		{name: "present_zero", raw: `{"type":"cli_exited","code":0}`, wantPresent: true, wantValue: 0},
